fixed moving ghost with 1 little bug
the ghost is moving quite well, using now a background buffer to restore, and a screen buffer to update the screen after calculation. Also, now, the ghost change his eyes, thanks to the change pos function, you'll see my cook.
BUT!!!! black pixel are randomly appearing after the ghost passage when it comes to the top of the screen only. very weird !