done the basic algorithm to load the maze from a table
we draw the maze in data, where each nibble = a 8x8 bloc of maze. the maze 's sprites are stored in MazeSprites , where each byte is a 256 color.
the function to draw the maze in the background buffer is 'MazeToBGbuffer'
we have to keep in mind the little endian in the drawing of the maze in the table (the model)